pyproject.toml
[project]
name = "QuickBBS"
version = "3.95.0"
authors = [
  { name="Benjamin Schollnick", email="Benjamin@schollnick.net" },
]
description = "Experimental Gallery / File Browser"
readme = "README.md"
requires-python = ">=3.12,<3.15"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dynamic = ["dependencies"]

[project.urls]
Homepage = "https://github.com/bschollnick/quickbbs"
Issues = "https://github.com/bschollnick/quickBBS/issues"

[tool.black]
line-length = 150

[tool.pyright]
exclude = ["**/gallery/**", "**/albums/**"]

[tool.pylint]
max-line-length = 150

[tool.pylance]
exclude = ["**/gallery/**", "**/albums/**"]

[tool.mypy]
python_version = "3.14"
explicit_package_bases = true
pretty = true
color_output = true
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]
exclude = ["albums", "gallery"]
mypy_path = "quickbbs"

[tool.django-stubs]
django_settings_module = "quickbbs.settings"

[tool.poetry]

[tool.poetry.group.dev.dependencies]
locust = "^2.42.0"
httpx = {extras = ["http2"], version = "^0.28.1"}

[tool.isort]
profile = "black"
src_paths = ["quickbbs"]
skip_gitignore = true
known_first_party = ["quickbbs", "frontend", "cache_watcher", "filetypes", "thumbnails"]

[tool.poetry.dependencies]
python = "^3.12,<3.15"
aiopath = "^0.7.7"
pymupdf = "^1.24.14"
django-icons = "^26"
django-htmx = "^1.21.0"
humanize = "^4.11.0"
django-jinja = "^2.11.0"
django-grappelli = "^4.0.1"
django-debug-toolbar = "^6.0.0"
djangorestframework = "^3.15.2"
django-filter = "^25.1"
django-allauth = "^65.2.0"
watchdog = "^6.0.0"
django-extensions = "^4.1"
django-ranged-fileresponse = {git = "https://github.com/devmonkey22/django-ranged-fileresponse"}
pillow = {url = "https://github.com/python-pillow/Pillow/archive/refs/heads/main.zip"}
markdown2 = "^2.5.1"
numpy = "^2.1.3"
django-thread = {git = "https://github.com/ross/django-thread"}
imagehash = "^4.3.1"
xattr = "^1.1.0"
dhash = "^1.4"
pathvalidate = "^3.2.1"
black = "^26"
isort = ">=6.0.0"
mypy = "^1.13.0"
natsort = ">=8.4.0"
psycopg = {extras = ["c"], version = "^3.2.3"}
certifi = "^2026"
gunicorn = ">=23.0.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.1"}
types-pyyaml = "^6.0.12.20240917"
uvicorn = ">=0.37.0"
pylint-django = ">=2.6.1"
hypercorn = ">=0.17.3"
pdfrw = "^0.4"
django-compression-middleware = "^0.5.0"
werkzeug = "^3.1.3"
pyopenssl = "^25.0.0"
msgspec = "^0.20"
more-itertools = "^10.6.0"
memoization = "^0.4.0"
cachetools = "^7"
django-db-connection-pool = "^1.2.6"
urllib3 = "^2.4.0"
cachecontrol = "^0.14.3"
aiofiles = ">=24.1.0"
pyobjc-core = "^12.0"
# pyobjc = "^12.0"
pyobjc-framework-uniformtypeidentifiers = "^12.0"
pyobjc-framework-cocoa = "^12.0"
ffmpeg-python = "^0.2.0"
charset-normalizer = "^3.4.2"
pytest = "^9.0.2"
pytest-django = "^4.11.1"
pytest-asyncio = "^1.3.0"
pyscan-rs = "^0.1.8"
pip-audit = "^2.9.0"
pipdeptree = "^2.28.0"
django = {version = "^6.0a1", allow-prereleases = true}
autoflake = "^2.3.1"
unidecode = "^1.4.0"
psycopg-pool = "^3.2.6"
django-dbbackup = "^5.0.0"
colorama = "^0.4.6"
pyobjc-framework-scriptingbridge = "^12.0"
pyobjc-framework-quartz = "^12.0"
pyobjc-framework-fsevents = "^12.0"
pyobjc-framework-coreservices = "^12.0"
pyobjc-framework-coremedia = "^12.0"
pyobjc-framework-coreaudio = "^12.0"
pyobjc-framework-avfoundation = "^12.0"
py-applescript = "^1.0.3"
httptools = "^0.7.1"
uvloop = "^0.22.1"
zensical = "^0.0.23"
djcheckup = "^0.6.0"
poetry = "^2.2.1"
django-ai-boost = "^0.7.0"
pylint = "^4.0.4"
wheel = "^0.46.3"
fido2 = "^2.1.1"
tprof = "^1.1.0"
django-dbtasks = "^0.4.0"
